+++ libc/ioctl/ioctl.c
@@ -3,9 +3,9 @@
 #include <sys/ioctl.h>
 #include <unistd.h>
 
-DEFN_SYSCALL3(ioctl, SYS_IOCTL, int, int, void *);
+DEFN_SYSCALL3(ioctl, SYS_IOCTL, int, unsigned long, void *);
 
-int ioctl(int fd, int request, void * argp) {
+int ioctl(int fd, unsigned long request, void * argp) {
 	return syscall_ioctl(fd, request, argp);
 }
